Fixed misc compil warnings

+++ src/isomedia/box_code_meta.c
@@ -245,7 +245,7 @@ GF_Err xml_Write(GF_Box *s, GF_BitStream *bs)
 	if (!s) return GF_BAD_PARAM;
 	e = gf_isom_full_box_write(s, bs);
 	if (e) return e;
-	if (ptr->xml) gf_bs_write_data(bs, ptr->xml, strlen(ptr->xml));
+	if (ptr->xml) gf_bs_write_data(bs, ptr->xml, (u32) strlen(ptr->xml));
 	gf_bs_write_u8(bs, 0);
 	return GF_OK;
 }
